+++ app/src/main/java/com/example/soundlink/features/auth/ui/screens/register/RegisterViewModel.kt
@@ -1,14 +1,11 @@
 package com.example.soundlink.features.auth.ui.screens.register
 
 import android.util.Log
-import android.widget.Toast
-import androidx.compose.ui.platform.LocalContext
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.soundlink.core.domain.model.User
-import com.example.soundlink.core.domain.repository.UserRepository
 import com.example.soundlink.core.ui.session.SessionViewModel
-import com.example.soundlink.features.auth.domain.usecases.RegisterUseCase
+import com.example.soundlink.features.auth.usecases.RegisterUseCase
 import kotlinx.coroutines.launch
 
 class RegisterViewModel(
@@ -17,7 +14,7 @@ class RegisterViewModel(
 ): ViewModel() {
 
 
-    fun register(name: String, email: String, pass: String, age: Long, onResult: (Boolean) -> Unit) {
+    fun register(name: String, email: String, pass: String, age: Int, onResult: (Boolean) -> Unit) {
 
         viewModelScope.launch{
             try{
